// frontend/src/types/generated/Page.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

export type Page = "home" | "shop" | "product" | "lot" | "cart" | "checkout" | "blog" | "blog_post" | "about_us" | "contact" | "cookie_policy" | "privacy_policy" | "newsletter" | "account" | "not_found";